+++ packages/common/src/data/reverse-router.ts
@@ -23,7 +23,7 @@ export function getReverseRouter(isMobile: boolean) {
       sterfte: () => `/landelijk/sterfte`,
       ziekenhuisopnames: () => `/landelijk/ziekenhuis-opnames`,
       intensiveCareOpnames: () => `/landelijk/intensive-care-opnames`,
-      verpleeghuiszorg: () => `/landelijk/verpleeghuiszorg`,
+      verpleeghuiszorg: () => `/landelijk/kwetsbare-groepen-70-plussers`,
       gehandicaptenzorg: () => `/landelijk/gehandicaptenzorg`,
       thuiswonendeOuderen: () => `/landelijk/thuiswonende-ouderen`,
       rioolwater: () => `/landelijk/rioolwater`,
@@ -36,44 +36,25 @@ export function getReverseRouter(isMobile: boolean) {
     },
 
     vr: {
-      index: (code?: string) =>
-        code
-          ? isMobile
-            ? `/veiligheidsregio/${code}`
-            : reverseRouter.vr.rioolwater(code)
-          : '/veiligheidsregio',
-      geldendeAdviezen: (code: string) =>
-        `/veiligheidsregio/${code}/geldende-adviezen`,
+      index: (code?: string) => (code ? (isMobile ? `/veiligheidsregio/${code}` : reverseRouter.vr.rioolwater(code)) : '/veiligheidsregio'),
+      geldendeAdviezen: (code: string) => `/veiligheidsregio/${code}/geldende-adviezen`,
       vaccinaties: (code: string) => `/veiligheidsregio/${code}/vaccinaties`,
-      positiefGetesteMensen: (code: string) =>
-        `/veiligheidsregio/${code}/positief-geteste-mensen`,
+      positiefGetesteMensen: (code: string) => `/veiligheidsregio/${code}/positief-geteste-mensen`,
       sterfte: (code: string) => `/veiligheidsregio/${code}/sterfte`,
-      ziekenhuisopnames: (code: string) =>
-        `/veiligheidsregio/${code}/ziekenhuis-opnames`,
-      verpleeghuiszorg: (code: string) =>
-        `/veiligheidsregio/${code}/verpleeghuiszorg`,
-      gehandicaptenzorg: (code: string) =>
-        `/veiligheidsregio/${code}/gehandicaptenzorg`,
-      thuiswonendeOuderen: (code: string) =>
-        `/veiligheidsregio/${code}/thuiswonende-ouderen`,
+      ziekenhuisopnames: (code: string) => `/veiligheidsregio/${code}/ziekenhuis-opnames`,
+      verpleeghuiszorg: (code: string) => `/veiligheidsregio/${code}/kwetsbare-groepen-70-plussers`,
+      gehandicaptenzorg: (code: string) => `/veiligheidsregio/${code}/gehandicaptenzorg`,
+      thuiswonendeOuderen: (code: string) => `/veiligheidsregio/${code}/thuiswonende-ouderen`,
       rioolwater: (code: string) => `/veiligheidsregio/${code}/rioolwater`,
       gedrag: (code: string) => `/veiligheidsregio/${code}/gedrag`,
-      brononderzoek: (code: string) =>
-        `/veiligheidsregio/${code}/brononderzoek`,
+      brononderzoek: (code: string) => `/veiligheidsregio/${code}/brononderzoek`,
     },
 
     gm: {
-      index: (code?: string) =>
-        code
-          ? isMobile
-            ? `/gemeente/${code}`
-            : reverseRouter.gm.rioolwater(code)
-          : '/gemeente',
-      positiefGetesteMensen: (code: string) =>
-        `/gemeente/${code}/positief-geteste-mensen`,
+      index: (code?: string) => (code ? (isMobile ? `/gemeente/${code}` : reverseRouter.gm.rioolwater(code)) : '/gemeente'),
+      positiefGetesteMensen: (code: string) => `/gemeente/${code}/positief-geteste-mensen`,
       sterfte: (code: string) => `/gemeente/${code}/sterfte`,
-      ziekenhuisopnames: (code: string) =>
-        `/gemeente/${code}/ziekenhuis-opnames`,
+      ziekenhuisopnames: (code: string) => `/gemeente/${code}/ziekenhuis-opnames`,
       rioolwater: (code: string) => `/gemeente/${code}/rioolwater`,
       vaccinaties: (code: string) => `/gemeente/${code}/vaccinaties`,
     },
